try to avoid move m3 video loader to common path
