+++ vllm/multimodal/video.py
@@ -1338,73 +1338,6 @@ class Qwen2VLVideoBackend(VideoBackend):
 
 
 @VIDEO_LOADER_REGISTRY.register(
-    name="minimax_m3_vl",
-    video_processor="MiniMaxM3VLVideoProcessor",
-)
-class MiniMaxM3VideoBackend(VideoBackend):
-    @classmethod
-    def load_bytes(
-        cls,
-        data: bytes,
-        num_frames: int = -1,
-        fps: int = 1,
-        max_duration: int = 300,
-        frame_recovery: bool = False,
-        *,
-        backend: Literal[
-            "opencv",
-            "pyav",
-            "torchcodec",
-            "pynvvideocodec",
-            "deepstream",
-        ] = "opencv",
-        **kwargs,
-    ) -> tuple[npt.NDArray, dict[str, Any]]:
-        return super().load_bytes(
-            data,
-            num_frames=num_frames,
-            fps=fps,
-            max_duration=max_duration,
-            frame_recovery=frame_recovery,
-            backend=backend,
-            **kwargs,
-        )
-
-    @classmethod
-    def compute_frames_index_to_sample(
-        cls,
-        source: VideoSourceMetadata,
-        target: VideoTargetMetadata,
-        **kwargs,
-    ) -> list[int]:
-        total_frames = source.total_frames_num
-        video_fps = source.original_fps
-        fps = target.fps
-        if total_frames <= 0 or video_fps <= 0 or fps <= 0:
-            return [0] if total_frames > 0 else []
-
-        read_time_interval = 1.0 / fps
-        eps = 1e-4
-
-        indices: list[int] = []
-        prev_kept_ts = -float("inf")
-        while True:
-            if not indices:
-                target_frame = 0
-            else:
-                target_ts = prev_kept_ts + read_time_interval - eps
-                target_frame = math.ceil(target_ts * video_fps)
-                target_frame = max(target_frame, indices[-1] + 1)
-            if target_frame >= total_frames:
-                break
-            indices.append(target_frame)
-            prev_kept_ts = target_frame / video_fps
-        if not indices:
-            indices = [0]
-        return indices
-
-
-@VIDEO_LOADER_REGISTRY.register(
     "opencv_dynamic",
     video_processor="Glm4vVideoProcessor",
 )
